requirements.txt
streamlit
boto3
pdfplumber
python-pptx
fpdf2
python-docx
